backend/requirements.txt
fastapi==0.109.0
# mysql_connector_repackaged==0.3.1
mysql-connector-python==8.3.0
pydantic==2.5.3
uvicorn==0.27.0
tensorflow==2.15.0
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.0
matplotlib==3.8.2
openai==1.10.0